Fixes to make sure the button is being disabled properly.
Watching the attribute value itself instead of the string the attribute value returns. Also, looking for the presence of ng-disabled to trigger the watch, since disabled won't be there unless the ng-disabled value is 'true'.

I have not changed anything to include the disabled attribute itself on the button, as the bootstrap class handles most of that, and I'm not totally sure what the implications would be.